[DNM] Exploring the code
The decomposition is fun, I think it just needs a slight tweak so that it becomes easy to listen to events emitted by the draggable components.  I experimentally rewrote some of the data wiring with this possibility in mind, but I'm too inexperienced with Vue to immediately see where to add the event handlers, so that you have a real-world API in which these handlers are invisible to the implementor (i.e. app.js).

Nice to debug out of the box, no source map is needed!